Refactor StreamConsumerFactory to use StreamMetadataProvider and StreamCOnsumer
